wifi_sniffer/websniffer/requirements.txt
flask==2.0.1
flask-socketio==5.1.1
flask-login==0.5.0
flask-sqlalchemy==3.1.1
scapy==2.4.5
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.3.1
plotly==5.3.1
dash==2.13.0
python-jose==3.3.0
weasyprint==60.1
dnspython==2.4.2
maxminddb==2.2.0
python-dotenv==0.19.0
werkzeug==2.0.1
eventlet==0.33.0
gevent==21.8.0
gevent-websocket==0.10.1
abuseipdb==0.1.0
virustotal-python==1.0.1
shodan==1.30.0 